feat: major refactor for code cleanliness and modularity
