+++ src/constants.js
@@ -13,3 +13,4 @@ export const SUN_DIAMETER = 1392684;
  */
 export const SCALE_FACTOR = 10000000;
 
+export const timeScale = { value: 1 };
